deprecate DummyPrintConnector in favour of MemoryPrintConnector (a rename w/ backwards compatibility)
