+++ doc/FAQ.md
@@ -10,7 +10,7 @@ The purpose of this driver it to generate binary ESC/POS code, which is understo
 
 If the printer understands ESC/POS, and you know how to get raw binary data to it, then yes. Otherwise, no.
 
-The [list of printers that are known to work](https://github.com/mike42/escpos-php/blob/development/README.md#printers) is crowd-sourced. We appreciate it when developers try out the driver, then [file information on the bug tracker](https://github.com/mike42/escpos-php/issues/new) with some information about which features worked on their model of printer.
+The [list of printers that are known to work](https://github.com/mike42/escpos-php/blob/development/README.md#printers) is crowdsourced. We appreciate it when developers try out the driver, then [file information on the bug tracker](https://github.com/mike42/escpos-php/issues/new) with some information about which features worked on their model of printer.
 
 To see how well your printer works, first check that it supports ESC/POS, then begin by attempting to send the text "Hello World" to your printer on the command-line, from the computer that will run PHP.
 
@@ -24,7 +24,7 @@ If you encounter garbage output when you try to print images or special characte
 
 ## I have a printer that does not understand ESC/POS. Can I use this driver?
 
-No. The purpose of this driver it to generate binary ESC/POS code. If your printer doesn't understand that, then this code wont be much use to you.
+No. The purpose of this driver it to generate binary ESC/POS code. If your printer doesn't understand that, then this code won't be much use to you.
 
 Some printers do have an emulation mode for ESC/POS. The vendor docs will tell if this is the case, and how to enable it.
 
@@ -46,13 +46,13 @@ The connectors are-
 
 - `FilePrintConnector` and `NetworkPrintConnector` directly use files or network sockets.
 - `WindowsPrintConnector` and `CupsPrintConnector` tie in with Windows and Unix system printing.
-- `DummyPrintConnector` does not connect to a real printer, and can be used to save ESC/POS receipts to a database, for example.
+- `MemoryPrintConnector` does not connect to a real printer, and can be used capture ESC/POS receipts for testing or audit, for example.
 
-At this point, you might find that the way you would like to print is not supported by escpos-php. You can post your printing command as a feature request on the issue tracker.
+You might find that the way you would like to print is not supported by escpos-php. You can post your printing command as a feature request on the issue tracker.
 
 Lastly, you may run in to the final common trap:
 
-4. Your PHP is not running with the same sort of permissions as your login account. Again, no amount of PHP code can fix this. For example, on LAMP, your `www-data` user needs to be in the `lp` group, while on WAMP, `Local Service` account may run in to problems. SELinux and firewalls are also worth a look.
+4. Your PHP script is not running with the same sort of permissions as your login account. Again, it's not possible for the PHP code itself to fix this. For example, on LAMP, your `www-data` user needs to be in the `lp` group, while on WAMP, `Local Service` account may run in to problems. SELinux and firewalls are also worth a look.
 
 When printing fails, you can expect a PHP Exception that explains what went wrong. They are all clues:
 
@@ -66,7 +66,7 @@ Please file a bug if you think that there is a specific situation which escpos-p
 
 ## Can I print over the network?
 
-Certainly, as long as your printer is available over the network.
+Certainly, as long as your printer is available over the network!
 
 - `NetworkPrintConnector` will speak directly to an Ethernet-connected printer on port 9100.
 
@@ -75,7 +75,7 @@ For USB or Serial printers, you need to install the printer on a computer and th
 - `WindowsPrintConnector` will connect to Windows shared printers from Windows or Linux (Linux users will need Samba).
 - `CupsPrintConnector` will connect to CUPS-shared printers from Linux or Mac.
 
-Always start by testing your shared printer setup outside of escpos-php. The examples linked to in the README are commented with some example commands to get you started. Typically, networks, firewalls and permissions need to be set up.
+Always start by testing your shared printer setup outside of escpos-php. The examples linked to in the README are commented with some example commands to get you started. Typically, networks, firewalls and permissions all need to be set up.
 
 Once you have a working command to send text to your printer (from the PHP server), you are ready to use escpos-php.
 
@@ -127,7 +127,7 @@ First, ensure you have the Imagick plugin loaded. The driver will avoid a slower
 
 Next, connect over a faster interface. Serial printers have a low bit-rate, and the printer spends a lot of time waiting for data. If you have USB or Ethernet, then use it (note: storing graphics to the printer memory is not currently implemented).
 
-Lastly, the printer will go faster if you use less pixels. Since images are two-dimensional, scaling down by 50% removes 75% of the pixels. The driver can then print at a half the density, so that your lower resolution image appears the same size when printed.
+Lastly, images are sent uncompressed in ESC/POS, and you can speed up the transfer if you use fewer pixels. Images are two-dimensional, and scaling down by 50% removes 75% of the pixels. The driver has an option to print at half the density, so that the lower resolution image appears the same size when printed.
 
 ## How can I get the status of the printer?
